web_app/requirements.txt
fastapi
uvicorn
pillow
requests
numpy
pydantic~=1.10.17
python-multipart